Test 0

+++ libs/graphic_functions.h
@@ -8,8 +8,8 @@
 
 void setColor(SDL_Renderer* renderer, const Color& color);
 SDL_Rect makeRect(const int& x, const int& y, const int& w, const int& h);
-void circle(SDL_Renderer* renderer, const Complex& center, const float& radius, const Color& color);
-void circleFill(SDL_Renderer* renderer, const Complex& center, const float& radius, const Color& color);
-void outlinedCircle(SDL_Renderer* renderer, const Complex& center, const Color& outline, const Color& filling, const float& radius);
+void circle(SDL_Renderer* renderer, const Complex& center, const int& radius, const Color& color);
+void circleFill(SDL_Renderer* renderer, const Complex& center, const int& radius, const Color& color);
+void outlinedCircle(SDL_Renderer* renderer, const Complex& center, const Color& outline, const Color& filling, const int& radius);
 
 #endif
